moves ResponsiveTable in its own shared component
