+++ app/windows/Storage/Components/StorageList.jsx
@@ -3,31 +3,7 @@ import { isEqual } from 'underscore'
 import { observer } from 'mobx-react'
 
 import StorageElement from './StorageElement'
-import styled from 'styled-components'
-
-import Table from '../../../components/Table'
-
-const ResponsiveTable = styled(Table)`
-  table-layout:fixed;
-  min-width: 700px;
-
-  th:nth-child(1) {
-    width: 35px;
-  }
-
-  th:nth-child(2) {
-    width: 400px;
-  }
-
-  th:nth-child(3) {
-    width: 85px;
-  }
-
-  th:nth-child(4) {
-    width: 100%;
-  }
-`;
-
+import ResponsiveTable from '../../../components/ResponsiveTable'
 
 @observer
 class StorageList extends React.Component {
